Extract playerInput function
